Add GOALS.md
I need to keep track of the thoughts for my current task I want to
accomplish to drive the next process in the CLI. Currently, I need to
create a new state, each time the CLI runs. That state should check for
the existence of a configuration file and use that, otherwise,
instantiate a new state, which can be saved as a new configuration file.